+++ common/src/commonMain/kotlin/com/blockstream/common/Constants.kt
@@ -58,4 +58,6 @@ object Urls {
         "https://help.blockstream.com/hc/en-us/articles/4406185830041-Why-is-my-Blockstream-Jade-not-connecting-over-Bluetooth-"
     const val LEDGER_SUPPORTED_ASSETS = "https://docs.blockstream.com/green/hww/hww-index.html#ledger-supported-assets"
     const val BLUETOOTH_PERMISSIONS = "https://developer.android.com/guide/topics/connectivity/bluetooth/permissions"
+    const val SECURITY_WATCH_ONLY = "https://blkstrm.com/watchonly"
+
 }