+++ compose/src/commonMain/kotlin/com/blockstream/compose/screens/overview/SecurityScreen.kt
@@ -19,7 +19,9 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.text.SpanStyle
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import blockstream_green.common.generated.resources.Res
 import blockstream_green.common.generated.resources.id_biometrics
@@ -28,16 +30,20 @@ import blockstream_green.common.generated.resources.id_connect_hardware_wallet
 import blockstream_green.common.generated.resources.id_firmware_update
 import blockstream_green.common.generated.resources.id_genuine_check
 import blockstream_green.common.generated.resources.id_hardware
+import blockstream_green.common.generated.resources.id_learn_more
 import blockstream_green.common.generated.resources.id_mobile
 import blockstream_green.common.generated.resources.id_pin
 import blockstream_green.common.generated.resources.id_recovery
 import blockstream_green.common.generated.resources.id_recovery_phrase
 import blockstream_green.common.generated.resources.id_security_level_
 import blockstream_green.common.generated.resources.id_unlock_method
+import blockstream_green.common.generated.resources.id_watchonly
+import blockstream_green.common.generated.resources.id_watchonly_description
 import blockstream_green.common.generated.resources.id_your_device
 import blockstream_green.common.generated.resources.id_your_jade
 import com.adamglin.PhosphorIcons
 import com.adamglin.phosphoricons.Regular
+import com.adamglin.phosphoricons.regular.Binoculars
 import com.adamglin.phosphoricons.regular.CaretRight
 import com.adamglin.phosphoricons.regular.Cpu
 import com.adamglin.phosphoricons.regular.Fingerprint
@@ -46,12 +52,14 @@ import com.adamglin.phosphoricons.regular.Password
 import com.adamglin.phosphoricons.regular.PlugsConnected
 import com.adamglin.phosphoricons.regular.SealCheck
 import com.adamglin.phosphoricons.regular.ShieldChevron
+import com.blockstream.common.Urls
 import com.blockstream.common.data.AlertType
 import com.blockstream.common.data.CredentialType
 import com.blockstream.common.data.GreenWallet
 import com.blockstream.common.data.MenuEntry
 import com.blockstream.common.data.MenuEntryList
 import com.blockstream.common.data.SetupArgs
+import com.blockstream.common.events.Events
 import com.blockstream.common.models.overview.SecurityViewModel
 import com.blockstream.common.models.overview.SecurityViewModel.LocalSideEffects
 import com.blockstream.common.models.overview.SecurityViewModelAbstract
@@ -67,14 +75,20 @@ import com.blockstream.compose.components.GreenCard
 import com.blockstream.compose.components.ListHeader
 import com.blockstream.compose.components.OnProgressStyle
 import com.blockstream.compose.components.Promo
+import com.blockstream.compose.screens.overview.components.WatchOnlyWalletDescription
+import com.blockstream.compose.theme.bodyMedium
 import com.blockstream.compose.theme.displaySmall
 import com.blockstream.compose.theme.green
 import com.blockstream.compose.theme.labelLarge
+import com.blockstream.compose.theme.md_theme_primary
+import com.blockstream.compose.theme.titleMedium
 import com.blockstream.compose.theme.titleSmall
 import com.blockstream.compose.theme.whiteMedium
 import com.blockstream.compose.utils.SetupScreen
 import com.blockstream.ui.components.GreenColumn
 import com.blockstream.ui.components.GreenRow
+import com.blockstream.ui.components.RichSpan
+import com.blockstream.ui.components.RichText
 import com.blockstream.ui.navigation.LocalInnerPadding
 import com.blockstream.ui.navigation.getResult
 import com.blockstream.ui.utils.bottom
@@ -139,35 +153,37 @@ fun SecurityScreen(viewModel: SecurityViewModelAbstract) {
                 .plus(PaddingValues(bottom = 80.dp + 24.dp))
         ) {
 
-            item {
-                GreenColumn(
-                    horizontalAlignment = Alignment.CenterHorizontally,
-                    modifier = Modifier.fillMaxWidth()
-                ) {
-                    Icon(
-                        imageVector = PhosphorIcons.Regular.ShieldChevron,
-                        contentDescription = null,
-                        modifier = Modifier.size(32.dp)
-                    )
-                    Text(
-                        text = stringResource(
-                            Res.string.id_security_level_,
-                            if (isHardware) "II" else "I"
-                        ), color = whiteMedium
-                    )
-                    Text(
-                        text = stringResource(if (isHardware) Res.string.id_hardware else Res.string.id_mobile),
-                        style = displaySmall
-                    )
+            if (!isWatchOnly) {
+                item {
+                    GreenColumn(
+                        horizontalAlignment = Alignment.CenterHorizontally,
+                        modifier = Modifier.fillMaxWidth()
+                    ) {
+                        Icon(
+                            imageVector = PhosphorIcons.Regular.ShieldChevron,
+                            contentDescription = null,
+                            modifier = Modifier.size(32.dp)
+                        )
+                        Text(
+                            text = stringResource(
+                                Res.string.id_security_level_,
+                                if (isHardware) "II" else "I"
+                            ), color = whiteMedium
+                        )
+                        Text(
+                            text = stringResource(if (isHardware) Res.string.id_hardware else Res.string.id_mobile),
+                            style = displaySmall
+                        )
 
-                    if (!isHardware) {
-                        GreenButton(
-                            text = stringResource(Res.string.id_compare_security_levels),
-                            type = GreenButtonType.OUTLINE,
-                            size = GreenButtonSize.BIG,
-                            color = GreenButtonColor.GREENER
-                        ) {
-                            viewModel.postEvent(NavigateDestinations.SecurityLevel(greenWallet = viewModel.greenWallet))
+                        if (!isHardware) {
+                            GreenButton(
+                                text = stringResource(Res.string.id_compare_security_levels),
+                                type = GreenButtonType.OUTLINE,
+                                size = GreenButtonSize.BIG,
+                                color = GreenButtonColor.GREENER
+                            ) {
+                                viewModel.postEvent(NavigateDestinations.SecurityLevel(greenWallet = viewModel.greenWallet))
+                            }
                         }
                     }
                 }
@@ -237,7 +253,26 @@ fun SecurityScreen(viewModel: SecurityViewModelAbstract) {
                 }
 
             } else if (isWatchOnly) {
+                item {
+                    GreenColumn(
+                        horizontalAlignment = Alignment.CenterHorizontally,
+                        modifier = Modifier.fillMaxWidth()
+                    ) {
+                        Icon(
+                            imageVector = PhosphorIcons.Regular.Binoculars,
+                            contentDescription = null,
+                            modifier = Modifier.size(32.dp)
+                        )
+                        
+                        Text(
+                            text = stringResource(Res.string.id_watchonly),
+                            style = titleMedium,
+                            modifier = Modifier.padding(top = 8.dp)
+                        )
 
+                        WatchOnlyWalletDescription { viewModel.postEvent(Events.OpenBrowser(Urls.SECURITY_WATCH_ONLY)) }
+                    }
+                }
             } else {
 
                 if (showRecoveryConfirmation) {